chore: drop 222: Inline attach sources provider
